Adição dos dashboards para main
